paging + search
Listings are now paginated, and the search is semi-functional.

The number of items displayed per page is determined by NUM_PER_PAGE in navigation.js

Right now, search is implemented, but not filtering by genre. Seems as if that attr isn't in the db yet, even though it's on the form. Shouldn't be hard to add tomorrow